build: Add a minimal configure.ac
When building with slibtool using the rlibtool symlink the build will
fail when it fails to find the generated libtool. This is required so
rlibtool can determine if it should build shared or static libraries.

This can be fixed by adding a minimal configure.ac that can generate the
required files with autoreconf.

Gentoo Bug: https://bugs.gentoo.org/828492