Do not stop searching on access denied.
Stopping on EPERM and EACCESS prevents nvim from finding the correct file when it is executed via sudoedit.

When running as sudoedit on NixOS, the first directory of TERMINFO_DIRS is /root/.nix-profile/share/terminfo, however the process is executed by non-root user. Even though the correct file doesn't exist in that directory, we get EPERM instead of ENOENT, because /root can't be listed by non-root users by default.

This causes the search to terminate and nvim ends up with builtin_ansi &term.